Junit Test Class changes and removal old reinforce and fortify classes
