+++ src/com/appriskgame/controller/MapOperations.java
@@ -1,4 +1,4 @@
-package com.appriskgame.services;
+package com.appriskgame.controller;
 
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
