+++ src/com/appriskgame/controller/MapValidation.java
@@ -1,4 +1,4 @@
-package com.appriskgame.services;
+package com.appriskgame.controller;
 
 import java.io.BufferedReader;
 import java.io.FileReader;
