+++ test/com/appriskgame/test/JUnitTestSuite.java
@@ -5,8 +5,7 @@ import org.junit.runners.Suite;
 import org.junit.runners.Suite.SuiteClasses;
 
 @RunWith(Suite.class)
-@SuiteClasses({ MapOperationTest.class, MapValidationTest.class, ReinforcementPhaseTest.class, StartupPhaseTest.class,
-		FortificationPhaseTest.class })
+@SuiteClasses({ MapOperationTest.class, MapValidationTest.class, StartupPhaseTest.class, PlayerTest.class })
 
 /**
  * A TestSuite class for testing all the test cases
